Update qdq configs
