+++ examples/clip/laion_CLIP-ViT-B-32-laion2B-s34B-b79K_ptq_qdq.json
@@ -82,7 +82,6 @@
             "op_types_to_quantize": [ "MatMul", "LayerNormalization", "Gemm", "Sigmoid", "Gelu" ],
             "activation_type": "QUInt16",
             "weight_type": "QUInt8",
-            "quant_format": "QDQ",
             "calibrate_method": "MinMax"
         }
     },
