+++ examples/resnet/README.md
@@ -76,7 +76,7 @@ python -m pip install -r requirements.txt
 
 ### Prepare data and model
 ```
-python prepare_model_data.py
+python prepare_data.py
 ```
 
 ### Run sample using config
