+++ examples/resnet/imagenet.py
@@ -9,6 +9,7 @@ import numpy as np
 import torchvision.transforms as transforms
 from torch import from_numpy
 from torch.utils.data import Dataset
+import transformers
 
 from olive.data.registry import Registry
 
@@ -24,13 +25,16 @@ class ImagenetDataset(Dataset):
         return min(len(self.images), len(self.labels))
 
     def __getitem__(self, idx):
-        return {"input": self.images[idx]}, self.labels[idx]
+        return {"pixel_values": self.images[idx]}, self.labels[idx]
 
 
 @Registry.register_post_process()
 def imagenet_post_fun(output):
-    return output.argmax(axis=1)
-
+    return (
+        output.logits.argmax(axis=1)
+        if isinstance(output, transformers.modeling_outputs.ModelOutput)
+        else output.argmax(axis=1)
+    )
 
 preprocess = transforms.Compose(
     [
