--- examples/resnet/resnet_dynamic_ptq_cpu.json
@@ -1,65 +1,0 @@
-{
-    "input_model": {
-        "type": "PyTorchModel",
-        "model_path": "models/resnet_trained_for_cifar10.pt",
-        "io_config": {
-            "input_names": [ "input" ],
-            "input_shapes": [ [ 1, 3, 32, 32 ] ],
-            "output_names": [ "output" ],
-            "dynamic_axes": { "input": { "0": "batch_size" }, "output": { "0": "batch_size" } }
-        }
-    },
-    "systems": {
-        "local_system": {
-            "type": "LocalSystem",
-            "accelerators": [ { "execution_providers": [ "CPUExecutionProvider" ] } ]
-        }
-    },
-    "data_configs": [
-        {
-            "name": "cifar10_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "cifar10_val_dataset", "data_dir": "data" },
-            "dataloader_config": { "batch_size": 16, "drop_last": true }
-        }
-    ],
-    "evaluators": {
-        "common_evaluator": {
-            "metrics": [
-                {
-                    "name": "accuracy",
-                    "type": "custom",
-                    "data_config": "cifar10_data_config",
-                    "sub_types": [
-                        {
-                            "name": "accuracy_custom",
-                            "priority": 1,
-                            "higher_is_better": true,
-                            "goal": { "type": "max-degradation", "value": 0.01 }
-                        }
-                    ],
-                    "user_config": { "user_script": "user_script.py", "evaluate_func": "eval_accuracy" }
-                },
-                {
-                    "name": "latency",
-                    "type": "latency",
-                    "data_config": "cifar10_data_config",
-                    "sub_types": [
-                        { "name": "avg", "priority": 2, "goal": { "type": "percent-min-improvement", "value": 20 } }
-                    ]
-                }
-            ]
-        }
-    },
-    "passes": {
-        "onnx_conversion": { "type": "OnnxConversion", "target_opset": 13 },
-        "onnx_dynamic_quantization": { "type": "OnnxDynamicQuantization", "weight_type": "QUInt8" },
-        "session_params_tuning": { "type": "OrtSessionParamsTuning", "data_config": "cifar10_data_config" }
-    },
-    "search_strategy": { "execution_order": "joint", "sampler": "sequential" },
-    "host": "local_system",
-    "target": "local_system",
-    "evaluator": "common_evaluator",
-    "cache_dir": "cache",
-    "output_dir": "models/resnet_dynamic_ptq_cpu"
-}
