+++ examples/resnet/resnet_multiple_ep.json
@@ -1,5 +1,14 @@
 {
-    "input_model": { "type": "ONNXModel", "model_path": "models/resnet_trained_for_cifar10.onnx" },
+    "input_model": {
+        "type": "HfModel",
+        "model_path": "microsoft/resnet-50",
+        "task": "image-classification",
+        "io_config": {
+            "input_names": [ "pixel_values" ],
+            "input_shapes": [ [ 1, 3, 224, 224 ] ],
+            "output_names": [ "logits" ]
+        }
+    },
     "systems": {
         "python_system": {
             "type": "PythonEnvironment",
@@ -12,23 +21,17 @@
     },
     "data_configs": [
         {
-            "name": "cifar10_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "cifar10_val_dataset", "data_dir": "data" },
-            "dataloader_config": { "batch_size": 1, "drop_last": true }
-        },
-        {
-            "name": "cifar10_accuracy_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "cifar10_val_dataset", "data_dir": "data" },
-            "dataloader_config": { "batch_size": 1, "drop_last": true },
-            "post_process_data_config": { "type": "cifar10_post_process" }
-        },
-        {
-            "name": "quant_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "simple_dataset" },
-            "dataloader_config": { "type": "resnet_calibration_dataloader", "data_dir": "data" }
+            "name": "data_config",
+            "type": "HuggingfaceContainer",
+            "user_script": "imagenet.py",
+            "load_dataset_config": {
+                "data_name": "imagenet-1k",
+                "split": "validation",
+                "streaming": true,
+                "trust_remote_code": true
+            },
+            "pre_process_data_config": { "type": "dataset_pre_process", "size": 256, "cache_key": "imagenet" },
+            "post_process_data_config": { "type": "imagenet_post_fun" }
         }
     ],
     "evaluators": {
@@ -37,20 +40,20 @@
                 {
                     "name": "accuracy",
                     "type": "accuracy",
-                    "data_config": "cifar10_accuracy_data_config",
+                    "data_config": "data_config",
                     "sub_types": [
                         {
                             "name": "accuracy_score",
                             "priority": 1,
-                            "metric_config": { "task": "multiclass", "num_classes": 1000 }
+                            "metric_config": { "task": "multiclass", "num_classes": 1001 }
                         }
                     ]
                 },
                 {
                     "name": "latency",
                     "type": "latency",
-                    "data_config": "cifar10_data_config",
-                    "sub_types": [ { "name": "avg", "priority": 2 } ]
+                    "data_config": "data_config",
+                    "sub_types": [ { "name": "avg" } ]
                 }
             ]
         }
@@ -58,11 +61,11 @@
     "passes": {
         "onnx_quantization": {
             "type": "OnnxQuantization",
-            "data_config": "quant_data_config",
+            "data_config": "data_config",
             "weight_type": "QUInt8",
             "activation_type": "QUInt8"
         },
-        "session_params_tuning": { "type": "OrtSessionParamsTuning", "data_config": "cifar10_data_config" }
+        "session_params_tuning": { "type": "OrtSessionParamsTuning", "data_config": "data_config" }
     },
     "evaluator": "common_evaluator",
     "target": "python_system",
