+++ examples/resnet/resnet_ptq_qdq.json
@@ -1,26 +1,27 @@
 {
     "input_model": {
-        "type": "PyTorchModel",
-        "model_path": "models/resnet_trained_for_cifar10.pt",
+        "type": "HfModel",
+        "model_path": "microsoft/resnet-50",
+        "task": "image-classification",
         "io_config": {
-            "input_names": [ "input" ],
-            "input_shapes": [ [ 1, 3, 32, 32 ] ],
-            "output_names": [ "output" ],
-            "dynamic_axes": { "input": { "0": "batch_size" }, "output": { "0": "batch_size" } }
+            "input_names": [ "pixel_values" ],
+            "input_shapes": [ [ 1, 3, 224, 224 ] ],
+            "output_names": [ "logits" ]
         }
     },
     "data_configs": [
         {
-            "name": "cifar10_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "cifar10_val_dataset", "data_dir": "data" },
-            "dataloader_config": { "batch_size": 16, "drop_last": true }
-        },
-        {
-            "name": "quant_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "simple_dataset" },
-            "dataloader_config": { "type": "resnet_calibration_dataloader", "data_dir": "data" }
+            "name": "quantize_data_config",
+            "type": "HuggingfaceContainer",
+            "user_script": "imagenet.py",
+            "load_dataset_config": {
+                "data_name": "imagenet-1k",
+                "split": "validation",
+                "streaming": true,
+                "trust_remote_code": true
+            },
+            "pre_process_data_config": { "type": "dataset_pre_process", "size": 256, "cache_key": "imagenet" },
+            "post_process_data_config": { "type": "imagenet_post_fun" }
         }
     ],
     "evaluators": {
@@ -28,45 +29,45 @@
             "metrics": [
                 {
                     "name": "accuracy",
-                    "type": "custom",
+                    "type": "accuracy",
+                    "data_config": "quantize_data_config",
                     "sub_types": [
                         {
-                            "name": "accuracy_custom",
+                            "name": "accuracy_score",
                             "priority": 1,
-                            "higher_is_better": true,
-                            "goal": { "type": "max-degradation", "value": 0.05 }
+                            "metric_config": { "task": "multiclass", "num_classes": 1001 }
                         }
-                    ],
-                    "user_config": {
-                        "user_script": "user_script.py",
-                        "evaluate_func": "eval_accuracy",
-                        "evaluate_func_kwargs": { "data_dir": "data", "batch_size": 16 }
-                    }
+                    ]
                 },
                 {
                     "name": "latency",
                     "type": "latency",
-                    "data_config": "cifar10_data_config",
-                    "sub_types": [
-                        { "name": "avg", "priority": 2, "goal": { "type": "percent-min-improvement", "value": 10 } }
-                    ]
+                    "data_config": "quantize_data_config",
+                    "sub_types": [ { "name": "avg" } ]
                 }
             ]
         }
     },
     "passes": {
-        "onnx_conversion": { "type": "OnnxConversion", "target_opset": 13 },
-        "onnx_quantization": {
+        "conversion": {
+            "device": "cpu",
+            "type": "OnnxConversion",
+            "target_opset": 17,
+            "save_as_external_data": true,
+            "all_tensors_to_one_file": true,
+            "dynamic": false,
+            "use_dynamo_exporter": false
+        },
+        "OnnxQuantization": {
             "type": "OnnxQuantization",
-            "data_config": "quant_data_config",
+            "data_config": "quantize_data_config",
+            "activation_type": "QUInt16",
             "weight_type": "QUInt8",
-            "quant_format": "QDQ",
-            "activation_type": "QUInt8",
-            "calibrate_method": "MinMax",
-            "quant_preprocess": true
+            "calibrate_method": "MinMax"
         }
     },
     "evaluator": "common_evaluator",
     "cache_dir": "cache",
+    "clean_cache": true,
     "output_dir": "models/resnet_ptq_qdq"
 }
