+++ examples/resnet/resnet_qat_default_train_loop_cpu.json
@@ -1,12 +1,12 @@
 {
     "input_model": {
-        "type": "PyTorchModel",
-        "model_path": "models/resnet_trained_for_cifar10.pt",
+        "type": "HfModel",
+        "model_path": "microsoft/resnet-50",
+        "task": "image-classification",
         "io_config": {
-            "input_names": [ "input" ],
-            "input_shapes": [ [ 1, 3, 32, 32 ] ],
-            "output_names": [ "output" ],
-            "dynamic_axes": { "input": { "0": "batch_size" }, "output": { "0": "batch_size" } }
+            "input_names": [ "pixel_values" ],
+            "input_shapes": [ [ 1, 3, 224, 224 ] ],
+            "output_names": [ "logits" ]
         }
     },
     "data_configs": [
