+++ examples/resnet/resnet_static_ptq_cpu.json
@@ -1,32 +1,27 @@
 {
     "input_model": {
-        "type": "PyTorchModel",
-        "model_path": "models/resnet_trained_for_cifar10.pt",
+        "type": "HfModel",
+        "model_path": "microsoft/resnet-50",
+        "task": "image-classification",
         "io_config": {
-            "input_names": [ "input" ],
-            "input_shapes": [ [ 1, 3, 32, 32 ] ],
-            "output_names": [ "output" ],
-            "dynamic_axes": { "input": { "0": "batch_size" }, "output": { "0": "batch_size" } }
-        }
-    },
-    "systems": {
-        "local_system": {
-            "type": "LocalSystem",
-            "accelerators": [ { "execution_providers": [ "CPUExecutionProvider" ] } ]
+            "input_names": [ "pixel_values" ],
+            "input_shapes": [ [ 1, 3, 224, 224 ] ],
+            "output_names": [ "logits" ]
         }
     },
     "data_configs": [
         {
-            "name": "session_params_tuning_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "cifar10_val_dataset", "data_dir": "data" },
-            "dataloader_config": { "batch_size": 1, "drop_last": true }
-        },
-        {
-            "name": "quant_data_config",
-            "user_script": "user_script.py",
-            "load_dataset_config": { "type": "simple_dataset" },
-            "dataloader_config": { "type": "resnet_calibration_dataloader", "data_dir": "data" }
+            "name": "data_config",
+            "type": "HuggingfaceContainer",
+            "user_script": "imagenet.py",
+            "load_dataset_config": {
+                "data_name": "imagenet-1k",
+                "split": "validation",
+                "streaming": true,
+                "trust_remote_code": true
+            },
+            "pre_process_data_config": { "type": "dataset_pre_process", "size": 256, "cache_key": "imagenet" },
+            "post_process_data_config": { "type": "imagenet_post_fun" }
         }
     ],
     "evaluators": {
@@ -34,40 +29,34 @@
             "metrics": [
                 {
                     "name": "accuracy",
-                    "type": "custom",
-                    "data_config": "cifar10_data_config",
+                    "type": "accuracy",
+                    "data_config": "data_config",
                     "sub_types": [
                         {
-                            "name": "accuracy_custom",
+                            "name": "accuracy_score",
                             "priority": 1,
-                            "higher_is_better": true,
-                            "goal": { "type": "max-degradation", "value": 0.01 }
+                            "metric_config": { "task": "multiclass", "num_classes": 1001 }
                         }
-                    ],
-                    "user_config": { "user_script": "user_script.py", "evaluate_func": "eval_accuracy" }
+                    ]
                 },
                 {
                     "name": "latency",
                     "type": "latency",
-                    "data_config": "cifar10_data_config",
-                    "sub_types": [
-                        { "name": "avg", "priority": 2, "goal": { "type": "percent-min-improvement", "value": 20 } }
-                    ]
+                    "data_config": "data_config",
+                    "sub_types": [ { "name": "avg" } ]
                 }
             ]
         }
     },
     "passes": {
         "onnx_conversion": { "type": "OnnxConversion", "target_opset": 13 },
-        "onnx_static_quantization": { "type": "OnnxStaticQuantization", "data_config": "quant_data_config" },
+        "onnx_static_quantization": { "type": "OnnxStaticQuantization", "data_config": "data_config" },
         "session_params_tuning": {
             "type": "OrtSessionParamsTuning",
-            "data_config": "session_params_tuning_data_config"
+            "data_config": "data_config"
         }
     },
     "search_strategy": { "execution_order": "joint", "sampler": "sequential" },
-    "host": "local_system",
-    "target": "local_system",
     "evaluator": "common_evaluator",
     "cache_dir": "cache",
     "output_dir": "models/resnet_static_ptq_cpu"
