+++ examples/vit/vit_qdq.json
@@ -62,7 +62,6 @@
             "data_config": "quantize_data_config",
             "activation_type": "QUInt16",
             "weight_type": "QUInt8",
-            "quant_format": "QDQ",
             "calibrate_method": "MinMax",
             "quant_preprocess": true
         }
